+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 - Removed outdated code and reformatted the utils module
 - Lots of documentation edits
 - Merged the 'rocks update' command into the 'rocks status' command
-- datacloud objects are now iterated over their entries rather than their columns
-- datacloud objects are now subscriptable, following the syntax of subscripting numpy arrays
+- datacloud catalogues are now DataCloudDataFrame objects, a pd.DataFrame subclass with extra capabilities (weighted average and plotting)
 - The diamalbedo catalogue once again has the preferred attribute
 
 # 1.2.3 - 2021-09-08